requirements.txt
openai
langchain
pytest
python-dotenv